Move to an Exception approach
Move from an Error approach to an Exception approach. Throw exceptions
that have to be catch, instead of displaying an error.

Changes in approach: A function have its normal behaviour, and throw an
exception if something bad happen.
- Only provide exceptions to create and throw
- "False" is the only semantical value to return errors.
- To retrieve the error type and message, the exception must be catch.

Changes:
- `sass-init` do not initalize errors options, but only `prefix` an
global variables.
- `sass-errors-throw` create an exception and save it globally. The
exception will be retrieved later.
- Add `@function sass-errors-new` (see below).

New functions behaviors: (?: optional parameter)
- `@function sass-errors-init($prefix?)`: initialize SE
- `@function sass-errors-new($error, $type?, $datas?, $prefix?)`:
create an Exception object and return it.
- `@function sass-errors-new($error, $type?, $datas?, $prefix?)`:
create an Exception object and throw it. If an Exception is passed as
first parameter, throw it. Always return false: this function should be
called with `@return`.