Add `se-try`, `se-catch` and `se-try-catch` functions
Functions:
- `@function sass-errors-try($func_name, $func_args: ())`: Call a
function with its arguments. If the function return false without
throwing Exceptions, throw an Exception.
- `@function sass-errors-catch($types...)`: If an Exception has been
throw, return it. If types are given, return the Exception only if its
type matches with one of the given types. Otherwise, return false.
- `@function sass-errors-try-catch($try_func_name, $try_func_args:
(),$catch_func_name: null, $catch_func_args: (), $types: ())`: Call a
_try_ function with its arguments. If the function return false or
throw an Exception, catch it and pass the Exception to the given
_catch_ function with its arguments.